bench: 3055770

+++ src/engine/search/search.cc
@@ -364,7 +364,8 @@ Score Search::PVSearch(int depth,
     // Reverse (Static) Futility Pruning: Cutoff if we think the position can't
     // fall below beta anytime soon
     if (depth <= 6 && eval < kMateScore - kMaxPlyFromRoot) {
-      const int futility_margin = depth * 75 - 50 * stack->improving_rate;
+      const int futility_margin =
+          depth * 75 - static_cast<int>(50 * stack->improving_rate);
       if (eval - futility_margin >= beta) {
         return eval;
       }
